Include list of updated status properties in SharingDeviceListener.update_device
In the current form, no information what has been updated is shared with SharingDeviceListener. This makes it impossible to implement event-based integrations as required by switches for instance as it is unclear which state the device reported if it didn't change. 

As an example, the device https://www.aliexpress.com/item/1005004644217026.html implements the following instruction set: https://developer.tuya.com/en/docs/iot/s?id=Kbeoa9fkv6brp

Without this change it is impossible to handle events on switch_mode1 as the device might keep reporting "click".